server/requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
googletrans==4.0.0rc1
Pillow==8.2.0
pytesseract==0.3.7
